# [SPARK-39851][SQL] Improve join stats estimation if one side can keep uniqueness